Revert "Revert "Delete training session impl""
This reverts commit 540d798c61bf40b29757eedd3a4dde13a9f5d0e7.

+++ src/app/components/app-users/app-users.component.ts
@@ -64,7 +64,7 @@ export class AppUsersComponent implements OnInit {
 
   deleteUser(appUser:AppUser)
   {
-    if(confirm("Delete user '"+appUser.name+" "+ appUser.surname+" ?")) {
+    if(confirm("Delete user '"+appUser.name+" "+ appUser.surname+" and all payments and attendances connected?")) {
       this.appUsersService.deleteUser(appUser).subscribe(response=>{
         this.loadGroup();
       })
